fix(lifecycle): ship portable merge proof
Consumer repositories must select merge strategy from exact committed ADR evidence without depending on this repository's CI-only script or fetching missing proof objects.

CHANGELOG: Ship a fail-closed installed ADR merge-method verifier for all supported governance hosts.

Ref: ADR-0034
