+++ .github/scripts/check_adr_lifecycle.py
@@ -7,7 +7,6 @@ import hashlib
 import json
 import os
 import re
-import subprocess
 import sys
 
 REPO = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
@@ -15,7 +14,8 @@ sys.path.insert(0, os.path.join(REPO, "core", "pysrc"))
 
 import adr_lifecycle as al
 import prepare_adr_acceptance as paa
-from _gitexec import root_bound_git_env
+import _adrlifecyclegit as lifecycle_git
+from _adrlifecyclegit import _git, _git_blob, _ledger_at, source_ancestry
 
 
 LEDGER_REL = ".codearbiter/decisions/adr-lifecycle.jsonl"
@@ -37,83 +37,6 @@ def select_base_ref(event_name, event):
     return value
 
 
-def _git(root, *args):
-    return subprocess.run(
-        ["git", "--no-replace-objects", "-C", root, *args],
-        capture_output=True, check=False,
-        env=root_bound_git_env(),
-    )
-
-
-def _git_blob(root, commit, path):
-    if not isinstance(commit, str):
-        return None
-    resolved = _git(root, "rev-parse", "--verify", "%s^{commit}" % commit)
-    if resolved.returncode != 0:
-        return None
-    result = _git(root, "show", "%s:%s" % (commit, path))
-    return result.stdout if result.returncode == 0 else None
-
-
-def _ledger_at(root, commit):
-    entry = _git(root, "ls-tree", "--name-only", commit, LEDGER_REL)
-    if entry.returncode != 0:
-        raise ValueError("could not inspect lifecycle ledger at %s" % commit)
-    if not entry.stdout.strip():
-        return None
-    blob = _git_blob(root, commit, LEDGER_REL)
-    if blob is None:
-        raise ValueError("could not read lifecycle ledger at %s" % commit)
-    return blob
-
-
-def source_ancestry(root, events, current_ref, base_ref=None):
-    """Require retained source identities; select a merge that preserves them."""
-    errors = []
-    refs = {}
-    for label, ref in (("current", current_ref), ("base", base_ref)):
-        if ref is None and label == "base":
-            continue
-        if not isinstance(ref, str) or not ref:
-            errors.append("%s ref is not a resolvable commit: %s" % (label, ref))
-            continue
-        resolved = _git(root, "rev-parse", "--verify", "--end-of-options",
-                        "%s^{commit}" % ref)
-        if resolved.returncode:
-            errors.append("%s ref is not a resolvable commit: %s" % (label, ref))
-        else:
-            refs[label] = resolved.stdout.decode("ascii").strip()
-    if errors:
-        return errors, None
-    method = "squash"
-    sources = set()
-    for event in events:
-        if not isinstance(event, dict):
-            continue
-        kind = event.get("event")
-        if kind in ("acceptance", "implemented", "verified"):
-            source = event.get("source_commit")
-        elif kind == "baseline":
-            source = event.get("observed_commit")
-        else:
-            continue
-        if not isinstance(source, str) or not al.HEX40.fullmatch(source):
-            errors.append("lifecycle source commit is malformed")
-            continue
-        sources.add(source)
-    for source in sorted(sources):
-        retained = _git(root, "merge-base", "--is-ancestor", source, refs["current"])
-        if retained.returncode == 1:
-            errors.append("%s: source commit is not an ancestor of current ref" % source)
-        elif retained.returncode:
-            errors.append("%s: could not verify source commit ancestry" % source)
-        if "base" in refs:
-            in_base = _git(root, "merge-base", "--is-ancestor", source, refs["base"])
-            if in_base.returncode == 1:
-                method = "merge"
-            elif in_base.returncode:
-                errors.append("%s: could not verify source commit base ancestry" % source)
-    return errors, None if errors else method
 
 
 def _index_blob(root, path):
@@ -511,30 +434,8 @@ def main(argv=None):
         # Explicit revision checks are derived from committed bytes, never a dirty
         # working-tree ledger that can hide a new source binding.
         try:
-            committed = _ledger_at(args.root, args.current_ref)
-            if committed is None:
-                raise ValueError("current ref has no lifecycle ledger")
-            events = [json.loads(line) for line in committed.decode("utf-8").splitlines()
-                      if line.strip()]
-            tree = _git(args.root, "ls-tree", "-r", "--name-only", "-z",
-                        args.current_ref, ".codearbiter/decisions/")
-            if tree.returncode:
-                raise ValueError("could not inspect current ADR tree")
-            blobs = {}
-            for path in tree.stdout.decode("utf-8").split("\0"):
-                prefix = ".codearbiter/decisions/"
-                if not path.startswith(prefix):
-                    continue
-                name = path[len(prefix):]
-                if "/" in name:
-                    continue
-                match = al.ADR_RE.fullmatch(name)
-                if match:
-                    blob = _git_blob(args.root, args.current_ref, path)
-                    if blob is None:
-                        raise ValueError("could not read current ADR blob: %s" % path)
-                    al.parse_adr(blob)
-                    blobs[match.group(1)] = blob
+            events, blobs, _committed = lifecycle_git.read_committed_evidence(
+                args.root, args.current_ref)
         except (ValueError, UnicodeError) as exc:
             errors.append("could not read committed current ref evidence: %s" % exc)
     else:
@@ -551,32 +452,15 @@ def main(argv=None):
             args.root, events, blobs, allow_local_pending=False)
     errors.extend(binding_errors)
 
-    source_blobs = {}
-    source_inputs = {}
-    for event in events:
-        if not isinstance(event, dict):
-            continue
-        kind = event.get("event")
-        if kind in ("acceptance", "baseline"):
-            commit = event.get("source_commit" if kind == "acceptance" else "observed_commit")
-            adr = event.get("adr")
-            if isinstance(commit, str) and isinstance(adr, str):
-                source_blobs[(commit, adr)] = _git_blob(
-                    args.root, commit, ".codearbiter/decisions/%s.md" % adr)
-        elif kind in ("implemented", "verified"):
-            commit = event.get("source_commit")
-            digests = event.get("input_digests")
-            if isinstance(digests, dict) and isinstance(commit, str):
-                for path in digests:
-                    if isinstance(path, str):
-                        source_inputs[(commit, path)] = _git_blob(args.root, commit, path)
-    errors.extend(al.validate_source_blobs(events, source_blobs))
-    errors.extend(al.validate_evidence_sources(events, source_inputs))
+    errors.extend(lifecycle_git.validate_committed_sources(args.root, events))
     merge_method = None
     if args.current_ref is not None:
-        ancestry_errors, merge_method = source_ancestry(
-            args.root, events, args.current_ref,
-            args.base_ref if args.merge_method else None)
+        if args.merge_method:
+            ancestry_errors, merge_method = lifecycle_git.merge_method(
+                args.root, args.base_ref, args.current_ref)
+        else:
+            ancestry_errors, merge_method = source_ancestry(
+                args.root, events, args.current_ref)
         errors.extend(ancestry_errors)
 
     if args.base_ref:
